move test to unit file
